pyproject.toml
[project]
name = "vk-photos"
version = "1.1.1"
description = ""
authors = [
    {name = "SolitarySpiral",email = "solitary.ratty@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles (==24.1.0)",
    "aiohttp (==3.11.18)",
    "pytils (==0.4.3)",
    "pytrovich (==0.0.4)",
    "pyyaml (==6.0.2)",
    "requests (==2.32.3)",
    "tqdm (==4.67.1)",
    "vk-api (==11.9.9)",
    "yt-dlp (==2025.4.30)",
    "asyncio (>=3.4.3,<4.0.0)",
    "pathlib (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
